fix() memory leakage with event handlers
fixed the memory leakage issues with event handlers as seen in issue 4.
Due to ws being an extension of "eventEmmiter" it was impossible to
create more then one event listener per event, which caused problems
when using .once in a function one might use a multiple times within a
short timespan (if the function is async (logically it should be (we are
talking about events here))).